+++ client/src/features/results/PlaceResults.tsx
@@ -12,6 +12,7 @@ import { getEstimates } from '../../lib/parse/search/searchResultParser'
 import { ResultsTab } from '../../types/ResultsTab'
 import StyledResultsCol from '../../styles/features/results/ResultsCol'
 import StyledEntityResultsRow from '../../styles/features/results/EntityResultsRow'
+import { DEFAULT_PAGE_LENGTH } from '../../config/searchTypes'
 
 import Paginate from './Paginate'
 import ResultsHeader from './ResultsHeader'
@@ -30,6 +31,9 @@ const PlaceResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
   const paramPrefix = getParamPrefix(tab)
   const pageParam = `${paramPrefix}p`
   const page: any = queryString.has(pageParam) ? queryString.get(pageParam) : 1
+  const pageLength: number = queryString.has('pageLength')
+    ? parseInt(queryString.get('pageLength')!, 10)
+    : DEFAULT_PAGE_LENGTH
   const sort = queryString.get(`${tab}Sort`)
   const view: string = queryString.has('view')
     ? (queryString.get('view') as string)
@@ -100,11 +104,11 @@ const PlaceResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
                     {resultsList(data.orderedItems)}
                   </Row>
                 )}
-                {estimate >= 20 && (
+                {estimate >= pageLength && (
                   <Paginate
                     estimate={estimate}
                     currentPage={parseInt(page, 10)}
-                    pageSize={20}
+                    pageSize={pageLength}
                   />
                 )}
               </React.Fragment>