+++ client/src/features/results/SetSnippet.tsx
@@ -18,6 +18,7 @@ import { getNextSetUris } from '../../lib/util/hierarchyHelpers'
 import config from '../../config/config'
 import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
 import theme from '../../styles/theme'
+import { DEFAULT_PAGE_LENGTH } from '../../config/searchTypes'
 
 import ProductionSnippet from './ProductionSnippet'
 import SnippetHeader from './SnippetHeader'
@@ -37,7 +38,7 @@ const SetSnippet: React.FC<ISearchData> = ({
   view,
   totalResults,
   index,
-  pageLength = 20,
+  pageLength = DEFAULT_PAGE_LENGTH,
   titleOfTabbedContent,
 }) => {
   const [isMobile, setIsMobile] = useState<boolean>(